pyproject.toml
[project]
name = "mercados"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.1.1",
    "langchain>=0.3.21",
    "python-dotenv>=1.1.0",
]
